Update create image to show it blocks
In the future I waill create a non-blocking call. Also note that the
docker API will push stdout through the connection, so a custome session
with a socket would be desireably he so we can strem from the socket as
the API pushes. Will have to look into this more.